🚨 tests: moving endings tests into separate files
