+++ tests/accido_test/pronoun_test.py
@@ -3,7 +3,6 @@ import sys
 
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "../..")))
 
-
 import pytest
 from src.core.accido.edge_cases import PRONOUNS
 from src.core.accido.endings import Noun, Pronoun
